Complete JV rewrite.
